refactor: Extract tool execution and token management utilities
Eliminate ~150 lines of duplicated code from chat endpoint by extracting
reusable helper functions into separate modules:

- token_management.py: Token counting and rolling window utilities
- tool_execution.py: Tool preparation, execution, and result conversion

This refactoring reduces chat_endpoints.py from 454 to 308 lines (32% reduction)
while improving code maintainability and reusability.
