+++ mcp_open_client/api/endpoints/conversations/chat_endpoints.py
@@ -2,86 +2,18 @@
 Chat endpoint for LLM interactions.
 """
 
-import asyncio
 import json
 import uuid
 from datetime import datetime
-from typing import Any, Dict, Optional
 
 from fastapi import HTTPException, status
 from openai import OpenAI
 
 from ...models.conversation import ConversationChatRequest, ConversationChatResponse
-
-# Import the shared SSE service from the sse module
-from ..sse import get_local_sse_service
 from . import router
 from .dependencies import conversation_manager, provider_manager, server_manager
-
-
-def inject_required_context_arguments(input_schema: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Inject two additional required arguments to a tool's input schema.
-
-    These arguments force the LLM to explain:
-    1. Why it's using the tool
-    2. What to do in case of error
-
-    Args:
-        input_schema: Original tool input schema
-
-    Returns:
-        Modified schema with additional required arguments
-    """
-    # Create a copy to avoid modifying the original
-    modified_schema = (
-        input_schema.copy() if input_schema else {"type": "object", "properties": {}}
-    )
-
-    # Ensure properties exists
-    if "properties" not in modified_schema:
-        modified_schema["properties"] = {}
-
-    # Add the two required context arguments
-    modified_schema["properties"]["tool_usage_reason"] = {
-        "type": "string",
-        "description": "¿Para qué estás usando esta herramienta? Explica brevemente el propósito y contexto de uso.",
-    }
-
-    modified_schema["properties"]["error_handling_plan"] = {
-        "type": "string",
-        "description": "¿Qué debes hacer en caso de error? Describe tu plan de contingencia si la herramienta falla.",
-    }
-
-    # Update required fields
-    if "required" not in modified_schema:
-        modified_schema["required"] = []
-
-    # Add to required if not already present
-    if "tool_usage_reason" not in modified_schema["required"]:
-        modified_schema["required"].append("tool_usage_reason")
-    if "error_handling_plan" not in modified_schema["required"]:
-        modified_schema["required"].append("error_handling_plan")
-
-    return modified_schema
-
-
-def filter_context_arguments(arguments: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Remove context arguments before calling the actual MCP server.
-
-    The MCP server doesn't expect these arguments - they're only for the LLM.
-
-    Args:
-        arguments: Tool arguments including context arguments
-
-    Returns:
-        Filtered arguments without context arguments
-    """
-    filtered = arguments.copy()
-    filtered.pop("tool_usage_reason", None)
-    filtered.pop("error_handling_plan", None)
-    return filtered
+from .token_management import apply_rolling_window, emit_token_update
+from .tool_execution import execute_single_tool_call, prepare_tools_for_llm
 
 
 @router.post(
@@ -153,46 +85,9 @@ async def conversation_chat(conversation_id: str, request: ConversationChatReque
         )
 
     # Get tool definitions from MCP servers and create mapping
-    tools_for_llm = []
-    tool_server_mapping = {}  # Maps tool_name -> server_id
-
-    if enabled_tools:
-        for enabled_tool in enabled_tools:
-            # Get the server
-            server = server_manager.get_server(enabled_tool.server_id)
-            if not server or server.status.value != "running":
-                continue
-
-            try:
-                # Get all tools from the server
-                server_tools = await server_manager.get_server_tools(
-                    enabled_tool.server_id
-                )
-                # Find the specific tool
-                for tool in server_tools:
-                    if tool.name == enabled_tool.tool_name:
-                        # Inject required context arguments into the schema
-                        modified_schema = inject_required_context_arguments(
-                            tool.input_schema or {}
-                        )
-
-                        # Convert to OpenAI function format
-                        tools_for_llm.append(
-                            {
-                                "type": "function",
-                                "function": {
-                                    "name": tool.name,
-                                    "description": tool.description or "",
-                                    "parameters": modified_schema,
-                                },
-                            }
-                        )
-                        # Map tool name to server ID for later execution
-                        tool_server_mapping[tool.name] = enabled_tool.server_id
-                        break
-            except Exception:
-                # Skip tools that fail to load
-                continue
+    tools_for_llm, tool_server_mapping = await prepare_tools_for_llm(
+        enabled_tools, server_manager
+    )
 
     # Get OpenAI client
     client = OpenAI(
@@ -239,49 +134,13 @@ async def conversation_chat(conversation_id: str, request: ConversationChatReque
         while iteration < max_iterations:
             iteration += 1
 
-            # 🔥 FIX: Apply token counting and rolling window BEFORE each LLM call
+            # Apply token counting and rolling window BEFORE each LLM call
             print(
                 f"[DEBUG] Applying token counting and rolling window for iteration {iteration}"
             )
-
-            # Use the conversation manager's token counter to apply rolling window
-            from mcp_open_client.core.conversations.storage import ConversationStorage
-            from mcp_open_client.core.conversations.token_counter import TokenCounter
-
-            # Create temporary instances for token counting
-            storage = ConversationStorage()
-            token_counter = TokenCounter()
-
-            # Load conversation to get current limits
-            conversation = storage.load(conversation_id)
-            if conversation:
-                # Apply rolling window with the current messages_for_llm
-                if conversation.max_tokens or conversation.max_messages:
-                    messages_for_llm, token_count = token_counter.apply_rolling_window(
-                        messages_for_llm,
-                        max_tokens=conversation.max_tokens,
-                        max_messages=conversation.max_messages,
-                        model=model_name,
-                    )
-                    print(
-                        f"[DEBUG] Rolling window applied: {len(messages_for_llm)} messages, {token_count} tokens"
-                    )
-                else:
-                    # Just count tokens without applying window
-                    token_count = token_counter.count_message_tokens(
-                        messages_for_llm
-                        + [{"role": "system", "content": system_prompt}],
-                        model_name,
-                    )
-                    print(f"[DEBUG] Token count: {token_count} tokens")
-
-                messages_in_context = len(messages_for_llm)
-            else:
-                print(
-                    f"[WARNING] Could not load conversation for token counting, using original messages"
-                )
-                token_count = len(messages_for_llm) * 50  # Rough estimate
-                messages_in_context = len(messages_for_llm)
+            messages_for_llm, token_count, messages_in_context = apply_rolling_window(
+                conversation_id, messages_for_llm, model_name
+            )
 
             # Update request_params with the potentially trimmed messages
             request_params["messages"] = [
@@ -340,157 +199,12 @@ async def conversation_chat(conversation_id: str, request: ConversationChatReque
                     }
                 )
 
-                # Emit SSE event for tool calls detected
+                # Execute each tool call using the extracted helper function
                 for tool_call in response_message.tool_calls:
-                    tool_name = tool_call.function.name
-                    tool_args = json.loads(tool_call.function.arguments)
-
-                    print(
-                        f"[DEBUG] Executing tool call: {tool_name} (id: {tool_call.id})"
+                    # Execute tool and get result (handles SSE events internally)
+                    tool_result = await execute_single_tool_call(
+                        tool_call, tool_server_mapping, server_manager, conversation_id
                     )
-
-                    # Emit tool call event (non-blocking, don't let SSE errors interrupt flow)
-                    try:
-                        sse_service = get_local_sse_service()
-                        await sse_service.emit_tool_call(
-                            conversation_id,
-                            {
-                                "tool_call_id": tool_call.id,
-                                "tool_name": tool_name,
-                                "arguments": tool_args,
-                                "assistant_content": response_message.content,
-                                "timestamp": timestamp,
-                            },
-                        )
-                    except Exception as sse_error:
-                        print(
-                            f"[DEBUG] SSE emit_tool_call failed (non-critical): {sse_error}"
-                        )
-
-                # Execute each tool call
-                for tool_call in response_message.tool_calls:
-                    tool_name = tool_call.function.name
-                    tool_args = json.loads(tool_call.function.arguments)
-
-                    # Find which server has this tool
-                    server_id = tool_server_mapping.get(tool_name)
-                    if not server_id:
-                        tool_result = json.dumps(
-                            {"error": f"Tool '{tool_name}' not found"}
-                        )
-                        print(f"[DEBUG] Tool '{tool_name}' not found in mapping")
-                        # Emit error event (non-blocking, don't let SSE errors interrupt flow)
-                        try:
-                            sse_service = get_local_sse_service()
-                            await sse_service.emit_tool_error(
-                                conversation_id,
-                                {
-                                    "tool_call_id": tool_call.id,
-                                    "tool_name": tool_name,
-                                    "error": f"Tool '{tool_name}' not found",
-                                    "timestamp": datetime.utcnow().isoformat() + "Z",
-                                },
-                            )
-                        except Exception as sse_error:
-                            print(
-                                f"[DEBUG] SSE emit_tool_error failed (non-critical): {sse_error}"
-                            )
-                    else:
-                        try:
-                            # Filter out context arguments before sending to MCP server
-                            filtered_args = filter_context_arguments(tool_args)
-
-                            print(
-                                f"[DEBUG] Calling tool '{tool_name}' on server '{server_id}' with args: {filtered_args}"
-                            )
-                            # Execute the tool with filtered arguments
-                            tool_result_raw = await server_manager.call_server_tool(
-                                server_id=server_id,
-                                tool_name=tool_name,
-                                arguments=filtered_args,
-                            )
-
-                            # Convert MCP result to serializable format
-                            if isinstance(tool_result_raw, list):
-                                # Handle list of TextContent objects
-                                result_text = ""
-                                for item in tool_result_raw:
-                                    if hasattr(item, "text"):
-                                        result_text += item.text
-                                    elif isinstance(item, dict) and "text" in item:
-                                        result_text += item["text"]
-                                    else:
-                                        result_text += str(item)
-                                tool_result = result_text
-                            elif hasattr(tool_result_raw, "content"):
-                                # Handle CallToolResult object with content attribute
-                                content = tool_result_raw.content
-                                if isinstance(content, list):
-                                    result_text = ""
-                                    for item in content:
-                                        if hasattr(item, "text"):
-                                            result_text += item.text
-                                        elif isinstance(item, dict) and "text" in item:
-                                            result_text += item["text"]
-                                        else:
-                                            result_text += str(item)
-                                    tool_result = result_text
-                                else:
-                                    tool_result = str(content)
-                            elif hasattr(tool_result_raw, "text"):
-                                # Single TextContent object
-                                tool_result = tool_result_raw.text
-                            elif isinstance(tool_result_raw, (str, int, float, bool)):
-                                tool_result = str(tool_result_raw)
-                            elif isinstance(tool_result_raw, dict):
-                                tool_result = json.dumps(tool_result_raw)
-                            else:
-                                tool_result = str(tool_result_raw)
-
-                            print(
-                                f"[DEBUG] Tool '{tool_name}' returned result (length={len(str(tool_result))})"
-                            )
-
-                            # Emit tool response event (non-blocking, don't let SSE errors interrupt flow)
-                            try:
-                                sse_service = get_local_sse_service()
-                                await sse_service.emit_tool_response(
-                                    conversation_id,
-                                    {
-                                        "tool_call_id": tool_call.id,
-                                        "tool_name": tool_name,
-                                        "result": tool_result,
-                                        "timestamp": datetime.utcnow().isoformat()
-                                        + "Z",
-                                    },
-                                )
-                            except Exception as sse_error:
-                                print(
-                                    f"[DEBUG] SSE emit_tool_response failed (non-critical): {sse_error}"
-                                )
-
-                        except Exception as e:
-                            tool_result = json.dumps({"error": str(e)})
-                            print(
-                                f"[DEBUG] Tool '{tool_name}' failed with error: {str(e)}"
-                            )
-                            # Emit error event (non-blocking, don't let SSE errors interrupt flow)
-                            try:
-                                sse_service = get_local_sse_service()
-                                await sse_service.emit_tool_error(
-                                    conversation_id,
-                                    {
-                                        "tool_call_id": tool_call.id,
-                                        "tool_name": tool_name,
-                                        "error": str(e),
-                                        "timestamp": datetime.utcnow().isoformat()
-                                        + "Z",
-                                    },
-                                )
-                            except Exception as sse_error:
-                                print(
-                                    f"[DEBUG] SSE emit_tool_error failed (non-critical): {sse_error}"
-                                )
 
                     # Save tool response message
                     tool_msg_id = f"msg-{uuid.uuid4().hex[:16]}"
@@ -503,7 +217,7 @@ async def conversation_chat(conversation_id: str, request: ConversationChatReque
                         message_id=tool_msg_id,
                         timestamp=timestamp,
                         tool_call_id=tool_call.id,
-                        name=tool_name,
+                        name=tool_call.function.name,
                     )
 
                     # Add tool response to conversation history
@@ -512,7 +226,7 @@ async def conversation_chat(conversation_id: str, request: ConversationChatReque
                             "role": "tool",
                             "content": tool_result,
                             "tool_call_id": tool_call.id,
-                            "name": tool_name,
+                            "name": tool_call.function.name,
                         }
                     )
 
@@ -526,29 +240,9 @@ async def conversation_chat(conversation_id: str, request: ConversationChatReque
                     )
 
                 # Emit token update after tool execution
-                try:
-                    from ...core.conversations.token_counter import TokenCounter
-
-                    token_counter = TokenCounter()
-                    current_tokens = token_counter.count_message_tokens(
-                        messages_for_llm, model_name
-                    )
-
-                    sse_service = get_local_sse_service()
-                    await sse_service.emit_token_update(
-                        conversation_id,
-                        {
-                            "token_count": current_tokens,
-                            "messages_in_context": len(messages_for_llm),
-                            "iteration": iteration,
-                            "timestamp": datetime.utcnow().isoformat() + "Z",
-                        },
-                    )
-                    print(f"[DEBUG] Emitted token update: {current_tokens} tokens")
-                except Exception as sse_error:
-                    print(
-                        f"[DEBUG] SSE emit_token_update failed (non-critical): {sse_error}"
-                    )
+                await emit_token_update(
+                    conversation_id, messages_for_llm, model_name, iteration
+                )
 
                 # Update request params for next iteration
                 request_params["messages"] = [
